Deep Speak, Fish Tongues, Triton Accent
Adds Deep Speak, a language for Triton with it's own font, and Scandinavian syllables.

Repurposes the unused lizard tongue into a fish tongue that changes spoken s & z in to sss & zzz. Gives it a sprite.

Adds an accent for Triton. Mostly based off Norwegian/Scandinavian, please looks at this and ask for adjustments. BCBM2222 on the discord asked for this, and sent in a file that I trimmed down and adjusted. Honestly, the tongue and the accent might be a bit too much. Let me know if we need to cut it.